CHPL_TARGET_ARCH -> CHPL_TARGET_CPU, CHPL_TARGET_MACHINE -> CHPL_TARGET_ARCH
