+++ doc/rst/platforms/omnipath.rst
@@ -21,7 +21,7 @@ Used the following environment variables:
     export CHPL_LAUNCHER='gasnetrun_psm'
     export CHPL_COMM_SUBSTRATE='psm'
     export CHPL_GASNET_SEGMENT='everything'
-    export CHPL_TARGET_ARCH='native'
+    export CHPL_TARGET_CPU='native'
 
 Using Slurm for Job Launch
 ++++++++++++++++++++++++++
