+++ make/compiler/Makefile.nvidia
@@ -25,6 +25,6 @@ COMP_GEN_LFLAGS +=-L$(CUDA_LIB_PATH) -lcudart
 
 RUNTIME_CUDA_CFLAGS = $(DEVICEEMU) -arch sm_13 --compiler-options -fno-strict-aliasing -DUNIX -O3 -DENABLE_GPU
 
-#TODO : add --gpu-architecture and/or --gpu-code based on CHPL_TARGET_ARCH
+#TODO : add --gpu-architecture and/or --gpu-code based on CHPL_TARGET_CPU
 #TODO : device FAST_IEEE_FLOAT float with --use_fast_math perhaps
