+++ runtime/Makefile.help
@@ -40,10 +40,10 @@ printstuff:
 	@echo $(CHPL_MAKE_HOME)/make/platform/Makefile.$(CHPL_MAKE_PLATFORM)
 	@echo CHPL_MAKE_MEM = $(CHPL_MAKE_MEM)
 	@echo CHPL_MAKE_HOST_TARGET = $(CHPL_MAKE_HOST_TARGET)
-	@echo CHPL_MAKE_TARGET_ARCH = $(CHPL_MAKE_TARGET_ARCH)
-	@echo CHPL_MAKE_RUNTIME_ARCH = $(CHPL_MAKE_RUNTIME_ARCH)
-	@echo CHPL_MAKE_TARGET_ARCH_FLAG = $(CHPL_MAKE_TARGET_ARCH_FLAG)
-	@echo CHPL_MAKE_TARGET_BACKEND_ARCH = $(CHPL_MAKE_TARGET_BACKEND_ARCH)
+	@echo CHPL_MAKE_TARGET_CPU = $(CHPL_MAKE_TARGET_CPU)
+	@echo CHPL_MAKE_RUNTIME_CPU = $(CHPL_MAKE_RUNTIME_CPU)
+	@echo CHPL_MAKE_TARGET_CPU_FLAG = $(CHPL_MAKE_TARGET_CPU_FLAG)
+	@echo CHPL_MAKE_TARGET_BACKEND_CPU = $(CHPL_MAKE_TARGET_BACKEND_CPU)
 	@echo CHPL_ORIG_TARGET_COMPILER = $(CHPL_ORIG_TARGET_COMPILER)
 	@echo CXX_STD = $(CXX_STD)
 	@echo RUNTIME_CXXFLAGS += $(RUNTIME_CXXFLAGS)
