+++ test/chplenv/chplconfig/correctness/chplconfig
@@ -4,7 +4,8 @@ CHPL_HOST_PLATFORM=darwin       # inline comment test
 CHPL_HOST_COMPILER=clang
 CHPL_TARGET_PLATFORM=darwin
 CHPL_TARGET_COMPILER=clang
-CHPL_TARGET_ARCH=native
+CHPL_TARGET_ARCH=x86_64
+CHPL_TARGET_CPU=native
 CHPL_LOCALE_MODEL=flat
 CHPL_COMM=none
 CHPL_COMM_SUBSTRATE=none
