+++ test/chplenv/printchplenv/printchplenv.chpl
@@ -2,6 +2,7 @@ writeln('CHPL_HOME: ', CHPL_HOME);
 writeln('CHPL_TARGET_PLATFORM: ', CHPL_TARGET_PLATFORM);
 writeln('CHPL_TARGET_COMPILER: ', CHPL_TARGET_COMPILER);
 writeln('CHPL_TARGET_ARCH: ', CHPL_TARGET_ARCH);
+writeln('CHPL_TARGET_CPU: ', CHPL_TARGET_ARCH);
 writeln('CHPL_LOCALE_MODEL: ', CHPL_LOCALE_MODEL);
 writeln('CHPL_COMM: ', CHPL_COMM);
 if CHPL_COMM != 'none' && CHPL_COMM != 'ugni' then {
