+++ test/compflags/albrecht/chplenv/chplenv.chpl
@@ -1,6 +1,7 @@
 writeln("CHPL_TARGET_PLATFORM=",CHPL_TARGET_PLATFORM);
 writeln("CHPL_TARGET_COMPILER=",CHPL_TARGET_COMPILER);
 writeln("CHPL_TARGET_ARCH=",CHPL_TARGET_ARCH);
+writeln("CHPL_TARGET_CPU=",CHPL_TARGET_ARCH);
 writeln("CHPL_LOCALE_MODEL=",CHPL_LOCALE_MODEL);
 writeln("CHPL_COMM=",CHPL_COMM);
 writeln("CHPL_COMM_SUBSTRATE=",CHPL_COMM_SUBSTRATE);
