+++ test/library/standard/Prefetch/prefetch.skipif
@@ -1,6 +1,6 @@
 # All 64-bit x86 has prefetch, but some 32-bit intel did not
 # so it is necessary to supply -march=native for 32-bit testing.
-# This could be achieved working through the CHPL_TARGET_ARCH
+# This could be achieved working through the CHPL_TARGET_CPU
 # mechanism to be more robust for different platforms.
 # In fact, this test works for me on 32-bit linux if CHPL_TARGET_ARCH=native
 # is set before testing begins...
