+++ test/release/examples/benchmarks/lulesh/lulesh.execenv
@@ -3,5 +3,5 @@
 # Guard pages are expensive on high core-count systems. See issue #7533
 
 import os
-if os.getenv('CHPL_TARGET_ARCH', '') in ['mic-knl']:
+if os.getenv('CHPL_TARGET_CPU', '') in ['mic-knl']:
     print('QT_GUARD_PAGES=false')
