+++ util/chplenv/__init__.py
@@ -3,8 +3,7 @@ __all__ = [
     'chpl_atomics',
     'chpl_aux_filesys',
     'chpl_bin_subdir',
-    'chpl_make',
-    'chpl_platform',
+    'chpl_cpu',
     'chpl_comm',
     'chpl_comm_segment',
     'chpl_comm_substrate',
@@ -15,8 +14,9 @@ __all__ = [
     'chpl_launcher',
     'chpl_llvm',
     'chpl_locale_model',
-    'chpl_machine',
+    'chpl_make',
     'chpl_mem',
+    'chpl_platform',
     'chpl_regexp',
     'chpl_tasks',
     'chpl_timers',
