+++ util/chplenv/chpl_comm_substrate.py
@@ -5,7 +5,7 @@ import sys
 chplenv_dir = os.path.dirname(__file__)
 sys.path.insert(0, os.path.abspath(chplenv_dir))
 
-import chpl_arch, chpl_comm, chpl_platform, overrides
+import chpl_comm, chpl_platform, overrides
 from utils import memoize
