+++ util/chplenv/chpl_gmp.py
@@ -5,7 +5,7 @@ import sys
 chplenv_dir = os.path.dirname(__file__)
 sys.path.insert(0, os.path.abspath(chplenv_dir))
 
-import chpl_3p_gmp_configs, chpl_arch, chpl_compiler, chpl_platform, overrides
+import chpl_3p_gmp_configs, chpl_compiler, chpl_platform, overrides
 from chpl_home_utils import get_chpl_third_party
 from utils import memoize
