+++ util/chplenv/third_party_utils.py
@@ -5,7 +5,7 @@ import sys
 chplenv_dir = os.path.dirname(__file__)
 sys.path.insert(0, os.path.abspath(chplenv_dir))
 
-import chpl_arch, chpl_machine, chpl_compiler
+import chpl_arch, chpl_cpu, chpl_compiler
 import chpl_lib_pic, chpl_locale_model, chpl_platform
 from chpl_home_utils import get_chpl_home, get_chpl_third_party, using_chapel_module
 from utils import error, memoize, run_command
@@ -16,11 +16,11 @@ from utils import error, memoize, run_command
 #
 @memoize
 def default_uniq_cfg_path():
-    arch_val = chpl_arch.get('target', map_to_compiler=True,
-                             get_lcd=using_chapel_module()).arch
+    cpu_val = chpl_cpu.get('target', map_to_compiler=True,
+                           get_lcd=using_chapel_module()).arch
     return '{0}-{1}-{2}-{3}-{4}'.format(chpl_platform.get('target'),
-                                        chpl_machine.get('target'),
-                                        arch_val,
+                                        chpl_arch.get('target'),
+                                        cpu_val,
                                         chpl_compiler.get('target'),
                                         chpl_lib_pic.get())
